+++ CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGE LOG
 ## V2.0 Beta 1 (Upcoming)
 
 * Upgraded to laravel 5
+* Renamed the rackspace "password" config to "apiKey"
 
 
 ## V1.0 RC 1 (Upcoming)